--- Client/Sources/WhooshingClient/Extensions/ChannelHandles+Sendable.swift
@@ -1,6 +1,0 @@
-import NIOExtras
-import NIOCore
-
-extension LengthFieldPrepender: @retroactive @unchecked Sendable {}
-
-extension ByteToMessageHandler: @retroactive @unchecked Sendable {}
