+++ Sources/WhooshingClient/Extensions/Basics.swift
