+++ Sources/WhooshingClient/Extensions/Channel+AddrInfo.swift
