+++ Sources/WhooshingClient/Extensions/ChannelHandles+Sendable.swift
