+++ Sources/WhooshingClient/Extensions/HTTPTypes+Codable.swift
