+++ Sources/WhooshingClient/HTTP/SendableDictionary.swift
