+++ Sources/WhooshingClient/HTTP/SendableStorage.swift
