+++ Sources/WhooshingClient/HTTP/WebURI.swift
